Split cli_test.dart into separate files.
It was getting pretty huge. I didn't want to stuff more tests in there
for CLI integration of analysis_options, so splitting it up into smaller
files first.

There are no meaningful changes here, just moving code around.
